+++ crates/cubed-application/src/use_cases/mod.rs
@@ -1,5 +1,6 @@
 pub mod create_server;
 pub mod delete_server;
+pub mod download_server_jar;
 pub mod init_file_system;
 pub mod reserve_port;
 pub mod restart_server;
@@ -9,6 +10,7 @@ pub mod stop_server;
 
 pub use create_server::{CreateServer, CreateServerInput};
 pub use delete_server::DeleteServer;
+pub use download_server_jar::DownloadServerJar;
 pub use init_file_system::InitFileSystem;
 pub use reserve_port::ReservePort;
 pub use restart_server::RestartServer;